requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
readability-lxml==0.8.1
ollama==0.2.1
supabase==2.5.0
python-dotenv==1.0.1
